convert bash models tests to pytests
