--- pipelines/tests/fable/test_fable_fit.sh
@@ -1,28 +1,0 @@
-#!/bin/bash
-
-# Check if the base directory is provided as an argument
-if [[ $# -ne 3 ]]; then
-	echo "Usage: $0 <base_dir> <disease> <location>"
-	echo "Received $# arguments: $@"
-	exit 1
-fi
-
-BASE_DIR="$1"
-disease="$2"
-location="$3"
-
-for epiweekly_flag in "" "--epiweekly"; do
-	python pipelines/fable/forecast_fable.py \
-		--disease "$disease" \
-		--loc "$location" \
-		--output-dir "$BASE_DIR/2024-12-21_forecasts" \
-		--run-date 2024-12-21 \
-		--n-training-days 90 \
-		--n-samples 500 \
-		$epiweekly_flag
-done
-
-if [ "$?" -ne 0 ]; then
-	echo "TEST-MODE FAIL: Forecasting/postprocessing pipeline failed"
-	exit 1
-fi
