+++ pipelines/tests/integration/test_pipeline_end_to_end.py
@@ -1,24 +1,25 @@
-import os
-import shutil
 import time
 from contextlib import contextmanager
-from math import isclose
 from pathlib import Path
 
 import polars as pl
 import pytest
 
-from pipelines.data import prep_data
 from pipelines.data.generate_test_data import (
     DEFAULT_DISEASES,
     DEFAULT_LOCATIONS,
-    REPORT_DATE,
-    make_forecast_data,
 )
-from pipelines.epiautogp import epiautogp_forecast_utils as epiautogp_utils
-from pipelines.epiautogp import forecast_epiautogp as epiautogp_module
-from pipelines.fable import forecast_fable as fable_module
-from pipelines.pyrenew_hew import forecast_pyrenew as pyrenew_module
+from pipelines.tests.integration.model_test_utils import (
+    EXCLUDE_LAST_N_DAYS,
+    FORECAST_DIR_NAME,
+    N_TRAINING_DAYS,
+    assert_model_outputs,
+    configure_data_mode,
+    model_batch_dir,
+    run_epiautogp,
+    run_fable,
+    run_pyrenew,
+)
 from pipelines.utils.common_utils import (
     create_prop_samples,
     make_figures_from_model_fit_dir,
@@ -27,10 +28,6 @@ from pipelines.utils.common_utils import (
 )
 from pipelines.utils.postprocess_forecast_batches import main as postprocess_batches
 
-FORECAST_DIR_NAME = f"{REPORT_DATE.isoformat()}_forecasts"
-N_TRAINING_DAYS = 42
-N_FORECAST_DAYS = 14
-EXCLUDE_LAST_N_DAYS = 1
 EXPECTED_MODELS = [
     "daily_fable_e_other",
     "epiweekly_fable_e_other",
@@ -40,20 +37,6 @@ EXPECTED_MODELS = [
     "prop_epiweekly_aggregated_pyrenew_e_epiweekly_fable_e_other",
     "prop_pyrenew_e_epiautogp_nssp_daily_other",
 ]
-MOCK_DATA_MODE = "mock"
-REAL_DATA_MODE = "real"
-
-
-def _normalize_pmf(weights: list[float]) -> list[float]:
-    total = sum(weights)
-    pmf = [weight / total for weight in weights]
-    assert isclose(sum(pmf), 1.0)
-    return pmf
-
-
-GENERATION_INTERVAL_PMF = _normalize_pmf([64, 23, 9, 3, 1])
-DELAY_PMF = _normalize_pmf([0, 2, 17, 24, 20, 14, 9, 6, 4, 2, 1, 1])
-RIGHT_TRUNCATION_PMF = _normalize_pmf([1, 0, 0, 0])
 
 
 @contextmanager
@@ -69,98 +52,6 @@ def _status_step(message: str):
     else:
         elapsed = time.perf_counter() - started_at
         print(f"[pipeline-e2e] {message} complete in {elapsed:.1f}s", flush=True)
-
-
-@pytest.fixture
-def pipeline_workspace(request, tmp_path, monkeypatch):
-    retained_dir = request.config.getoption("--e2e-output-dir")
-    force = request.config.getoption("--e2e-force")
-
-    if retained_dir is None:
-        workspace = tmp_path / "pipeline-e2e"
-    else:
-        workspace = Path(retained_dir)
-        if workspace.exists():
-            if not force:
-                pytest.fail(
-                    f"Retained workspace already exists: {workspace}. "
-                    "Pass --e2e-force to remove it before running."
-                )
-            shutil.rmtree(workspace)
-
-    workspace.mkdir(parents=True, exist_ok=True)
-
-    monkeypatch.setenv("MPLCONFIGDIR", str(workspace / ".matplotlib"))
-    monkeypatch.setenv(
-        "XLA_FLAGS",
-        os.environ.get("XLA_FLAGS", "--xla_force_host_platform_device_count=2"),
-    )
-    return workspace
-
-
-def _run_fable(
-    workspace: Path, disease: str, location: str, *, epiweekly: bool = False
-) -> None:
-    fable_module.main(
-        disease=disease,
-        loc=location,
-        output_dir=workspace / FORECAST_DIR_NAME,
-        n_training_days=N_TRAINING_DAYS,
-        n_forecast_days=N_FORECAST_DAYS,
-        exclude_last_n_days=EXCLUDE_LAST_N_DAYS,
-        n_samples=40,
-        run_date=REPORT_DATE,
-        epiweekly=epiweekly,
-    )
-
-
-def _run_pyrenew(workspace: Path, disease: str, location: str) -> None:
-    pyrenew_module.main(
-        disease=disease,
-        loc=location,
-        priors_path=Path("pipelines/pyrenew_hew/priors/prod_priors.py"),
-        output_dir=workspace / FORECAST_DIR_NAME,
-        n_training_days=N_TRAINING_DAYS,
-        n_forecast_days=N_FORECAST_DAYS,
-        exclude_last_n_days=EXCLUDE_LAST_N_DAYS,
-        n_chains=1,
-        n_samples=40,
-        n_warmup=40,
-        run_date=REPORT_DATE,
-        rng_key=12345,
-        fit_ed_visits=True,
-        forecast_ed_visits=True,
-    )
-
-
-def _run_epiautogp(workspace: Path, disease: str, location: str) -> None:
-    epiautogp_module.main(
-        disease=disease,
-        run_date=REPORT_DATE,
-        loc=location,
-        output_dir=workspace / FORECAST_DIR_NAME,
-        n_training_days=N_TRAINING_DAYS,
-        n_forecast_days=N_FORECAST_DAYS,
-        exclude_last_n_days=EXCLUDE_LAST_N_DAYS,
-        target="nssp",
-        frequency="daily",
-        ed_visit_type="other",
-        n_particles=2,
-        n_mcmc=2,
-        n_hmc=2,
-        n_forecast_draws=40,
-        smc_data_proportion=0.1,
-        n_threads=2,
-    )
-
-
-def _model_batch_dir(workspace: Path, disease: str) -> Path:
-    candidates = list((workspace / FORECAST_DIR_NAME).glob(f"{disease.lower()}_r_*"))
-    assert len(candidates) == 1, (
-        f"Expected one batch directory for {disease}, "
-        f"found {len(candidates)} in {workspace / FORECAST_DIR_NAME}"
-    )
-    return candidates[0]
 
 
 def _run_fusions(model_run_dir: Path) -> None:
@@ -203,114 +94,38 @@ def _run_fusions(model_run_dir: Path) -> None:
         model_fit_dir_to_hub_tbl(fusion_model_dir)
 
 
-def _assert_model_outputs(model_run_dir: Path) -> None:
-    for model_name in EXPECTED_MODELS:
-        model_dir = model_run_dir / model_name
-        assert model_dir.is_dir(), f"Missing model directory: {model_dir}"
-        assert (model_dir / "samples.parquet").is_file(), (
-            f"Missing samples parquet: {model_dir}"
-        )
-        assert (model_dir / "hubverse_table.parquet").is_file(), (
-            f"Missing hubverse table: {model_dir}"
-        )
-
-
-def _patch_dataops(monkeypatch) -> None:
-    def load_forecast_data(
-        *,
-        disease,
-        loc_abb,
-        run_date,
-        first_training_date,
-        last_training_date,
-        sources,
-        **kwargs,
-    ):
-        return make_forecast_data(
-            location=loc_abb,
-            disease=disease,
-            sources=sources,
-            first_training_date=first_training_date,
-            last_training_date=last_training_date,
-        )
-
-    for module in (fable_module, pyrenew_module):
-        monkeypatch.setattr(module, "load_forecast_data", load_forecast_data)
-    monkeypatch.setattr(epiautogp_utils, "load_forecast_data", load_forecast_data)
-    monkeypatch.setattr(
-        prep_data,
-        "get_nnh_generation_interval_pmf",
-        lambda **kwargs: GENERATION_INTERVAL_PMF.copy(),
-    )
-    monkeypatch.setattr(
-        prep_data,
-        "get_nnh_delay_pmf",
-        lambda **kwargs: DELAY_PMF.copy(),
-    )
-    monkeypatch.setattr(
-        prep_data,
-        "get_nnh_right_truncation_pmf",
-        lambda **kwargs: RIGHT_TRUNCATION_PMF.copy(),
-    )
-    monkeypatch.setattr(
-        epiautogp_utils,
-        "get_nnh_right_truncation_pmf",
-        lambda **kwargs: RIGHT_TRUNCATION_PMF.copy(),
-    )
-
-
-def _has_external_dataops_env() -> bool:
-    try:
-        from cfa.cloudops.util import check_ext_env
-    except ImportError:
-        print(
-            "[pipeline-e2e] cfa.cloudops.util.check_ext_env is unavailable; "
-            "using mocked DataOps data.",
-            flush=True,
-        )
-        return False
-
-    return check_ext_env()
-
-
-def _data_mode(request) -> str:
-    requested_mode = request.config.getoption("--e2e-data-mode")
-    if requested_mode == "auto":
-        return REAL_DATA_MODE if _has_external_dataops_env() else MOCK_DATA_MODE
-    return requested_mode
-
-
 @pytest.mark.pipeline_e2e
 def test_reduced_pipeline_end_to_end(pipeline_workspace, monkeypatch, request):
     workspace = pipeline_workspace
-    data_mode = _data_mode(request)
-    print(f"[pipeline-e2e] Using {data_mode} DataOps data.", flush=True)
-
-    if data_mode == MOCK_DATA_MODE:
-        _patch_dataops(monkeypatch)
+    configure_data_mode(request, monkeypatch)
 
     for disease in DEFAULT_DISEASES:
         for location in DEFAULT_LOCATIONS:
             with _status_step(f"Running Fable for {disease}, {location}"):
-                _run_fable(workspace, disease, location)
+                run_fable(workspace, disease, location)
 
             with _status_step(f"Running epiweekly Fable for {disease}, {location}"):
-                _run_fable(workspace, disease, location, epiweekly=True)
+                run_fable(workspace, disease, location, epiweekly=True)
 
             with _status_step(f"Running PyRenew for {disease}, {location}"):
-                _run_pyrenew(workspace, disease, location)
+                run_pyrenew(workspace, disease, location)
 
             with _status_step(f"Running EpiAutoGP for {disease}, {location}"):
-                _run_epiautogp(workspace, disease, location)
+                run_epiautogp(workspace, disease, location)
 
             model_run_dir = (
-                _model_batch_dir(workspace, disease) / "model_runs" / location
+                model_batch_dir(workspace, disease) / "model_runs" / location
             )
             with _status_step(f"Running fusion models for {disease}, {location}"):
                 _run_fusions(model_run_dir)
 
             with _status_step(f"Checking model outputs for {disease}, {location}"):
-                _assert_model_outputs(model_run_dir)
+                assert_model_outputs(
+                    workspace,
+                    disease,
+                    location,
+                    EXPECTED_MODELS,
+                )
 
     with _status_step("Postprocessing forecast batches"):
         postprocess_batches(
@@ -322,7 +137,7 @@ def test_reduced_pipeline_end_to_end(pipeline_workspace, monkeypatch, request):
 
     for disease in DEFAULT_DISEASES:
         with _status_step(f"Checking postprocessed outputs for {disease}"):
-            batch_dir = _model_batch_dir(workspace, disease)
+            batch_dir = model_batch_dir(workspace, disease)
             batch_info = parse_model_batch_dir_name(batch_dir.name)
             postprocessed_path = (
                 batch_dir
